# use the bias as intended