# Pull Request

## Related issue
Fixes https://github.com/meilisearch/arroy/issues/81

## What does this PR do?
- After looking into annoy it looks like we ignored the bias for the euclidean and manhattan distance, but using the bias the recall improved a little bit on small dimension datasets:
On main:
![image](https://github.com/user-attachments/assets/5cc0939d-e9a6-4d7b-8932-49e74ed7efb7)
After this PR:
![image](https://github.com/user-attachments/assets/6e210d51-8e9c-4ed8-9336-6b71004758fd)

> ![WARNING]
> Currently, this doesn't apply to the angular distance, we need to read the annoy code further to understand what they’re doing with the norm and if we’re doing everything correctly


### And for the binary quantization?

Currently, I don't really know how we should implements that for the binary quantization. I did a few tests here and some improved the recall sometimes for almost 10 points:
![image](https://github.com/user-attachments/assets/739ebf1c-5c95-4844-9bf9-f2771ee184e9)