@nnethercott 2025-05-30
Yo @irevoire I revived this PR in [a fork](https://github.com/meilisearch/arroy/compare/main...nnethercott:arroy:add-header-to-normal) cause it's concerning mathematically that the bias still isn't being used in arroy. 

I'll drop my thoughts below (déso pour le pavé)

TL;DR: for large datasets with moderate `n_trees` performance gaps are big. If you use tons of trees though you can generally overcome that gap (which is what arroy currently does so it might be OK)

# understanding the bias term
When the `arroy::Writer` is indexing, it builds separating hyperplanes to split leafs into two clusters. First `two_means` [finds the centroids](https://github.com/meilisearch/arroy/blob/main/src/distance/euclidean.rs#L54), $`\{c_{0}, c_{1}\}`$ and then we define a plane with unit normal, $`n`$, such that $`n \cdot c_{0}>0`$ and $`n \cdot c_{1}<0`$ 

A [hyperplane](https://en.wikipedia.org/wiki/Hyperplane#Vector_hyperplanes) is a set of points satisfying $`n \cdot x+b=0`$. We said earlier the plane we build passes through the midpoint of the centroids, so to find the bias we just plug that in and get 
```math
\begin{align*}
n \cdot (c_{0} + c_{1})/2 + b &= 0 \\ 
\implies - n \cdot (c_{0} + c_{1})/2 &= b
\end{align*}
```

That's the same as [what's in arroy](https://github.com/meilisearch/arroy/blob/main/src/distance/euclidean.rs#L68) (but currently its ignored)

Here's what the plane **should** look like (bias = $`- n \cdot (c_{0} + c_{1})/2`$): 
![scene](https://github.com/user-attachments/assets/299d6ddd-443b-4672-9d08-1779275de7c0)

Here's what it **currently** looks like (bias=0):
![scene_no_bias](https://github.com/user-attachments/assets/4e8ac5a3-1652-4879-8592-84010a17bc31)

## consequences
- the margin for two points belonging to separate clusters can be the same
- we potentially get poor splits at every iteration since the hyperplane doesn't necessarily cut the leaf subset in half
  -  leads to substantially longer indexing times due to more recursive calls until [fit_in_descendant](https://github.com/meilisearch/arroy/blob/main/src/writer.rs#L465) evaluates to true
 - pointless computation in `two_means` since we're only keeping angular info, the "seperating" plane we build may put all elements on the same side !
   - random projections are totally valid as a hash, we could accomplish this with less compute than what we're doing here 

> [!NOTE]  
> This is generally OK for Cosine and Dot distances since the data is normalized to the unit circle before indexing so the bias is close to 0 anyways. We're more worried about Euclidean and Manhattan...

# results 
## impact on recall@k
We'll use the [vector store benchmark](https://github.com/meilisearch/vector-store-relevancy-benchmark) for this. In the case where `|dataset|>>n_trees` using the bias generally outperforms the status quo.  We purposely limit the number of trees to assess the quality of learned `SplitPlaneNormal`s. 

### `|dataset|=100_000, n_trees = 1`

<details>
<summary>Click here</summary>

On main: 
![main_100k_1](https://github.com/user-attachments/assets/5bf9ab88-f7fe-4d34-b484-a21517cae85b)

After this PR:
![pr_100k_1](https://github.com/user-attachments/assets/32267e10-1bb5-4c69-8ed5-480ad506231c)
</details>

### `|dataset|=100_000, n_trees = 100`
<details>
<summary>Click here</summary>

On main: 
![main_100k_100](https://github.com/user-attachments/assets/c12d00cb-c5d1-4f36-bf7c-c623fb35f8fc)

After this PR:
![pr_100k_100](https://github.com/user-attachments/assets/9be59b7c-4399-43fc-be80-772322ad28e7)

</details>

### `|dataset|=10_000, n_trees = 1`
<details>
<summary>Click here</summary>

On main:
![main_10k_1](https://github.com/user-attachments/assets/f71bdaaf-e42f-4a6b-b7f3-0df634c15ebe)

After this PR:
![pr_10k_1](https://github.com/user-attachments/assets/9748ab84-743d-46b8-8cbc-140f7027c58e)

</details>

### `|dataset|=10_000, n_trees = 100`
<details>
<summary>Click here</summary>

On main:
![main_10k_100](https://github.com/user-attachments/assets/489985dc-fe5c-4e78-8e1d-cfd3c68a6621)

After this PR:
![pr_10k_100](https://github.com/user-attachments/assets/1f735403-538b-40f2-85f2-e9cfd3742033)

</details>

## impact on indexing time 
We'll just look at one dataset and the Euclidean distance since the idea is more or less the same for other datasets/distances. We expect main to be slower since imbalanced splits (e.g. those arising from not using the bias) require more recursive branching.

Here are the results for indexing 100000 vectors of dimension 768:

|      | n_trees=1 | n_trees=10 | n_trees=100 | n_trees=500 |
|------|-----------|------------|-------------|-------------|
| main | 2.08s     | 5.93s      | 55.94s      | 649.31s     |
| pr   | 3.44s     | 5.12s      | 48.82s      | 225.18s     |

# a weird thing 
If we run the benchmark and use the default number of trees, the approach with the proper use of the bias actually underperforms what's in arroy currently: 

On main: 
![main_10k_768](https://github.com/user-attachments/assets/6e4c8dd0-9ea1-4937-93c9-0097e0a21654)

After this PR:
![pr_10k_768](https://github.com/user-attachments/assets/b90a9655-8821-4b38-86f1-045c3bf1a22d)

My guess is that arroy overcomes poor quality trees with brute force search (high default number of trees, high oversampling) so the weak learners can ensemble better than the trees we produced by using the bias correctly. 

> [!WARNING] 
> The [`target_n_trees`](https://github.com/meilisearch/arroy/blob/main/src/writer.rs#L1231) is proprtional to the embedding dimension, irrespective of the number of vectors being indexed. 

from the code: 
```rust
// 1. How many descendants do we need:
let descendant_required = item_indices.len() / dimensions;
// 2. Find the number of tree nodes required per trees
let tree_nodes_per_tree = descendant_required + 1;
// 3. Find the number of tree required to get as many tree nodes as item:
let mut nb_trees = item_indices.len() / tree_nodes_per_tree;
```
In the above `nb_trees` = n/((n/d)+1) = d/(1+d/n) ≈ d for **any** collection we're indexing. So for my test above with 10k vectors we're using nearly 1k trees by default. This might explain the performance difference. 

# conclusions 
- for large datasets using the bias generally yields significant performance improvements in terms of recall and indexing time
- In cases where increasing `n_trees` is not feasible, using the bias properly should create higher quality trees
- Performance gains depend on dataset & embeddings structure

Very curious to hear your thoughts on this